+++ index.js
@@ -74,9 +74,9 @@ function parse(string) {
     throw new TypeError('argument string is required')
   }
 
-  // support req-like objects as argument
-  if (typeof string === 'object' && typeof string.headers === 'object') {
-    string = string.headers && string.headers['content-type']
+  // support req/res-like objects as argument
+  if (typeof string === 'object') {
+    string = getcontenttype(string)
   }
 
   if (typeof string !== 'string') {
@@ -113,6 +113,26 @@ function parse(string) {
   obj.parameters = params
 
   return obj
+}
+
+/**
+ * Get content-type from req/res objects.
+ *
+ * @param {object}
+ * @return {Object}
+ * @api private
+ */
+
+function getcontenttype(obj) {
+  if (typeof obj.getHeader === 'function') {
+    // res-like
+    return obj.getHeader('content-type')
+  }
+
+  if (typeof obj.headers === 'object') {
+    // req-like
+    return obj.headers && obj.headers['content-type']
+  }
 }
 
 /**