+++ HISTORY.md
@@ -2,6 +2,7 @@ unreleased
 ==========
 
   * Accept `req` as argument to `parse`
+  * Accept `res` as argument to `parse`
 
 0.0.0 / 2014-06-13
 ==================